+++ lib/numbers_and_words/helper_classes/array_extensions/helpers.rb
@@ -42,6 +42,10 @@ module NumbersAndWords
         self[0].to_i if self[0].to_i.positive?
       end
 
+      def only_ones
+        ones if !tens && !hundreds
+      end
+
       def teens
         tens_with_ones if tens == 1
       end