+++ lib/numbers_and_words/strategies/figures_converter/languages/nl.rb
@@ -10,9 +10,16 @@ module NumbersAndWords
           end
 
           def capacity_iteration
-            if FiguresArray::THOUSAND_CAPACITY == @current_capacity
-              capacity_words = words_in_capacity(@current_capacity)
-              capacity_words.empty? ? [] : [capacity_words, megs].join
+            return super if FiguresArray::THOUSAND_CAPACITY != @current_capacity
+            return megs if figures.number_in_capacity(@current_capacity) == 1
+
+            capacity_words = words_in_capacity(@current_capacity)
+            capacity_words.empty? ? [] : [capacity_words, megs].join
+          end
+
+          def simple_number_to_words
+            if @figures.only_ones && @figures.ones == 1
+              [@translations.t(:loose_one)]
             else
               super
             end