+++ lib/numbers_and_words/strategies/figures_converter/languages/pl.rb
@@ -1,25 +1,25 @@
+# frozen_string_literal: true
+
 module NumbersAndWords
-    module Strategies
-      module FiguresConverter
-        module Languages
-          class Pl < Base
+  module Strategies
+    module FiguresConverter
+      module Languages
+        class Pl < Base
+          def megs
+            super({ number: @figures.number_in_capacity(@current_capacity) })
+          end
 
-            def megs
-              super({:number => @figures.number_in_capacity(@current_capacity)})
-            end
+          private
 
-            private
+          def one?
+            words_in_capacity(@current_capacity) == [translations.ones(1)]
+          end
 
-            def one?
-              words_in_capacity(@current_capacity) == [translations.ones(1)]
-            end
-
-            def thousand?
-              FiguresArray::THOUSAND_CAPACITY == @current_capacity
-            end
-
+          def thousand?
+            FiguresArray::THOUSAND_CAPACITY == @current_capacity
           end
         end
       end
     end
   end
+end