+++ lib/numbers_and_words/strategies/figures_converter/decorators/pl.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require 'numbers_and_words/strategies/figures_converter/decorators/pl/base'
+require 'numbers_and_words/strategies/figures_converter/decorators/pl/fractional'
+require 'numbers_and_words/strategies/figures_converter/decorators/pl/integral'