+++ lib/numbers_and_words/translations/pl.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module NumbersAndWords
+  module Translations
+    class Pl < Base
+      include NumbersAndWords::Translations::Families::Latin
+      include NumbersAndWords::Translations::Extensions::FractionSignificance
+    end
+  end
+end