+++ spec/numbers_and_words/i18n/pluralization/pl_spec.rb
@@ -1,0 +1,24 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+describe I18n, '#pl' do
+  it 'pluralizes correctly' do
+    hash = {}
+    %w[one few many other].each { |key| hash[key.to_sym] = key }
+
+    expect(described_class.backend.send(:pluralize, :pl, hash, 1)).to eq('one')
+    expect(described_class.backend.send(:pluralize, :pl, hash, 2)).to eq('few')
+    expect(described_class.backend.send(:pluralize, :pl, hash, 3)).to eq('few')
+    expect(described_class.backend.send(:pluralize, :pl, hash, 5)).to eq('many')
+    expect(described_class.backend.send(:pluralize, :pl, hash, 10)).to eq('many')
+    expect(described_class.backend.send(:pluralize, :pl, hash, 11)).to eq('many')
+    expect(described_class.backend.send(:pluralize, :pl, hash, 21)).to eq('one')
+    expect(described_class.backend.send(:pluralize, :pl, hash, 29)).to eq('many')
+    expect(described_class.backend.send(:pluralize, :pl, hash, 52)).to eq('few')
+    expect(described_class.backend.send(:pluralize, :pl, hash, 131)).to eq('one')
+    expect(described_class.backend.send(:pluralize, :pl, hash, 1.31)).to eq('other')
+    expect(described_class.backend.send(:pluralize, :pl, hash, 2.31)).to eq('other')
+    expect(described_class.backend.send(:pluralize, :pl, hash, 3.31)).to eq('other')
+  end
+end