+++ lib/numbers_and_words/i18n/pluralization.rb
@@ -3,7 +3,6 @@
 require 'numbers_and_words/i18n/plurals/cs'
 require 'numbers_and_words/i18n/plurals/ru'
 require 'numbers_and_words/i18n/plurals/ua'
-require 'numbers_and_words/i18n/plurals/fr'
 require 'numbers_and_words/i18n/plurals/lv'
 require 'numbers_and_words/i18n/plurals/lt'