+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Features
   * Add support for Ruby 3.1. \[[#187](https://github.com/kslazarev/numbers_and_words/pull/187)\]
+  * Drop support for Ruby 2.5. \[[#180](https://github.com/kslazarev/numbers_and_words/pull/180)\]
   * Your contribution here.
 
 ## 0.11.11 (July 5, 2021)